Move container-state migration below the CLI layer
The session-rebuild consolidation gives the podman backend a
SessionResources collaborator that owns the destructive half of a
rebuild, including the legacy-state salvage that runs before a
container is replaced. That salvage currently lives in
cli/upgrade_persistence.py, so a backend calling it would invert the
dependency direction (backends -> cli), the same trap REFACTOR-005
records for SSH transport construction.

Move it down before anything depends on it:

- cli/upgrade_persistence.py -> backends/podman/legacy_state.py. None
  of persistent_state_paths, migrate_legacy_state, or
  credential_exclude_globs is CLI knowledge -- they drive a
  ContainerRunner and read agent config. Two lazy importers
  (cli/upgrade.py, cli/backup.py) and the test module follow.
- called_process_stderr() from cli/helpers.py to subprocess_utils.py,
  which legacy_state needs and which already owns the rest of the
  subprocess failure-translation helpers. Three other CLI importers
  follow; cli/helpers.py sheds its now-unused subprocess import.

Pure moves: no call site changes behaviour, and nothing patches either
symbol by path, so a missed import would be an ImportError at
collection rather than a silent pass. src/paude/backends,
src/paude/container and src/paude/transport now import nothing from
src/paude/cli.
